requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
telethon>=1.34.0
google-generativeai>=0.3.0
